backend/requirements.txt
|Django==4.2.8
djangorestframework==3.14.0
drf-spectacular==0.26.5
djangorestframework-simplejwt==5.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-json-logger==2.0.7
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
structlog==23.1.0
gunicorn==21.2.0
requests==2.31.0
